Fixed HttpStatus for Exceptions for Topic, User, Answers

+++ server/src/main/java/com/ivantanin/questionbase/service/UserService.java
@@ -87,8 +87,9 @@ public class UserService {
     @Transactional
     public void delete(String username) {
         User user = userRepository.findByUsername(username).orElse(null);
-        Long userId;
-        userId = user.getId();
+        if (user == null)
+            throw new NullPointerException();
+        Long userId = user.getId();
         answerRepository.deleteAllByUserId(userId);
         userRepository.deleteByUsername(username);
     }
